requirements.txt
python-telegram-bot
fastapi
uvicorn
python-dotenv
requests
pydantic